\name{dmr.util-package}
\alias{dmr.util-package}
\alias{dmr.util}

\docType{package}

\title{A package with utility functions used in the book Cichosz, P. (2015): Data Mining Algorithms: Explained Using R}

\description{A package with utility functions used in examples presented in the book Cichosz, P. (2015): Data Mining Algorithms: Explained Using R. See Appendix B or http://www.wiley.com/go/data_mining_algorithms for more details.}

\details{This package is a collection of very simple utility functions that are used in examples presented in most chapters of the book Cichosz, P. (2015): Data Mining Algorithms: Explained Using R. These include functions for basic calculations, data manipulation, and R formula manipulation that come handy when illustrating the operation of data mining algorithms.
}

\author{
Pawel Cichosz <p.cichosz@elka.pw.edu.pl>
}

\references{Cichosz, P. (2015): Data Mining Algorithms: Explained Using R. Wiley.}

\keyword{package}
\keyword{arith}
\keyword{univar}
\keyword{multivariate}

\seealso{
}

\examples{
}
